smart-vision-api/pyproject.toml
[build-system]
requires = ["setuptools>=70.2.0", "wheel>=0.45.1"]
build-backend = "setuptools.build_meta"

[project]
name = "smart_vision_api"
version = "1.0.0"
description = "FastAPI service for Smart Vision machine learning models and predictions"
readme = "README.md"
requires-python = ">=3.12.3"
license = {file = "LICENSE"}
authors = [
    {name = "SurplusGLOBAL AI Team"}
]
dependencies = [
    "fastapi>=0.115.12",
    "pydantic_settings==2.9.1",
    "FlagEmbedding>=1.3.4",
    "transformers==4.51.3",
    "pymilvus==2.5.10",
    "pymilvus[model]",
    "pymilvus[bulk_writer]",
    "python-multipart==0.0.20",
    "requests>=2.32.0",
    "beautifulsoup4>=4.12.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0"
]


[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "flake8>=7.2.0",
    "pytest>=8.3.5"
]

[tool.setuptools.packages.find]
include = ["smart_vision_api*"]

[tool.black]
line-length = 88
target-version = ["py312"]
